requirements.txt
transformers>=4.36.0
torch>=2.1.0
mergekit>=0.0.4
peft>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface-hub>=0.19.0
sentencepiece>=0.1.99
protobuf>=3.20.0
wandb>=0.16.0
einops>=0.7.0
safetensors>=0.4.0
llm-jp-eval>=1.0.0
auto-gptq>=0.7.0
awq>=0.1.0
gguf>=0.6.0